refactor: 아티스트 신청 조회 시, 장르까지 fetch join 하도록 쿼리 수정
